Refactoring MidiPlayer to a component folder
