+++ web/src/MidiPlayer/index.js
@@ -1,0 +1,3 @@
+import MidiPlayer from "./MidiPlayer";
+
+export default MidiPlayer;
